# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Swagger Json",
    "description": "This is a generated swagger json for your sails project",
    "termsOfService": "http://example.com/terms",
    "contact": {
      "name": "Theophilus Omoregbee",
      "url": "http://github.com/theo4u",
      "email": "theo4u@ymail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:1337"
    }
  ],
...
  "paths": {
    "/dashboard": {
      "get": {
        "summary": "View dashboard",
        "description": "Display \"Dashboard\" page.",
        "tags": [
          "View dashboard"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/csrfToken": {
      "get": {
        "summary": "/csrfToken",
        "tags": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "The requested resource"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/wallet/verifica": {
      "get": {
        "summary": "View verifica",
        "description": "Display \"Verifica\" page.",
        "tags": [
          "View verifica"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "initWallet",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "If true, the wallet will be initialized"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/v1/get-transaction": {
      "get": {
        "summary": "Get transaction by digest",
        "description": "Recupera una transazione dalla cache locale BlockchainData tramite digest.",
        "tags": [
          "Get transaction by digest"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "digest",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Transaction digest (hash) sulla blockchain IOTA"
          },
          {
            "in": "query",
            "name": "accountName",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "transactionId",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The requested resource"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/add-organizzazione": {
      "post": {
        "summary": "Add organizzazione",
        "description": "Aggiunge una nuova organizzazione al database e pubblica su blockchain.",
        "tags": [
          "Add organizzazione"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "denominazione",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 255,
              "minLength": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "string",
                      "description": "Organizzazione aggiunta con successo."
                    },
                    {
                      "type": "string",
                      "description": "I dati forniti non sono validi."
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/add-struttura": {
      "post": {
        "summary": "Add struttura",
        "description": "Aggiunge una nuova struttura e pubblica su blockchain.",
        "tags": [
          "Add struttura"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "denominazione",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 255,
              "minLength": 2
            }
          },
          {
            "in": "query",
            "name": "indirizzo",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2
            }
          },
          {
            "in": "query",
            "name": "organizzazione",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "in": "query",
            "name": "attiva",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "example": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "string",
                      "description": "Struttura aggiunta con successo."
                    },
                    {
                      "type": "string",
                      "description": "I dati forniti non sono validi."
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/add-lista": {
      "post": {
        "summary": "Add lista",
        "description": "Aggiunge una nuova lista al database e pubblica su blockchain.",
        "tags": [
          "Add lista"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "denominazione",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 255,
              "minLength": 2
            }
          },
          {
            "in": "query",
            "name": "struttura",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "string",
                      "description": "Lista aggiunta con successo."
                    },
                    {
                      "type": "string",
                      "description": "I dati forniti non sono validi."
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/add-assistito": {
      "post": {
        "summary": "Add assistito",
        "description": "Crea un assistito e pubblica su blockchain.",
        "tags": [
          "Add assistito"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "nome",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cognome",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "codiceFiscale",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z]{6}[0-9]{2}[A-Za-z][0-9]{2}[A-Za-z][0-9]{3}[A-Za-z]$"
            }
          },
          {
            "in": "query",
            "name": "dataNascita",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "email",
            "required": false,
            "schema": {
              "type": "string",
              "format": "email"
            }
          },
          {
            "in": "query",
            "name": "telefono",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "indirizzo",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "string",
                      "description": "Assistito aggiunto con successo."
                    },
                    {
                      "type": "string",
                      "description": "Success"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Add assistito",
        "description": "Crea un assistito e pubblica su blockchain.",
        "tags": [
          "Add assistito"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "nome",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cognome",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "codiceFiscale",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z]{6}[0-9]{2}[A-Za-z][0-9]{2}[A-Za-z][0-9]{3}[A-Za-z]$"
            }
          },
          {
            "in": "query",
            "name": "dataNascita",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "email",
            "required": false,
            "schema": {
              "type": "string",
              "format": "email"
            }
          },
          {
            "in": "query",
            "name": "telefono",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "indirizzo",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "string",
                      "description": "Assistito aggiunto con successo."
                    },
                    {
                      "type": "string",
                      "description": "Success"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/fetch-db-from-blockchain": {
      "post": {
        "summary": "Aggiorna db da blockchain",
        "description": "Aggiorna il database locale con i dati presenti nella cache blockchain.",
        "tags": [
          "Aggiorna db da blockchain"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "All done."
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/add-assistito-in-lista": {
      "post": {
        "summary": "Add assistito in lista",
        "description": "Aggiunge un assistito in una lista.",
        "tags": [
          "Add assistito in lista"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idAssistito",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "in": "query",
            "name": "idLista",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "string",
                      "description": "Assistito aggiunto in lista con successo."
                    },
                    {
                      "type": "string",
                      "description": "I dati forniti non sono validi"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/recover-from-arweave": {
      "post": {
        "summary": "Recover from Arweave",
        "description": "Recupera i dati dalla blockchain Arweave come fallback.",
        "tags": [
          "Recover from Arweave"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "dataType",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Tipo di dato da recuperare (MAIN_DATA, ASSISTITI_DATA, etc.). Se vuoto, recupera tutto."
          },
          {
            "in": "query",
            "name": "entityId",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "ID entita specifica da recuperare."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "string",
                      "description": "Dati recuperati con successo da Arweave."
                    },
                    {
                      "type": "string",
                      "description": "Arweave non configurato o errore nel recupero."
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/liste-dettaglio": {
      "get": {
        "summary": "API Liste Dettaglio",
        "description": "Ritorna il dettaglio di una lista con assistiti in coda (con posizione) e storico movimenti.",
        "tags": [
          "API Liste Dettaglio"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idLista",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dettaglio lista."
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/rimuovi-assistito-da-lista": {
      "post": {
        "summary": "Rimuovi assistito da lista",
        "description": "Rimuove un assistito dalla lista con gestione multi-lista.",
        "tags": [
          "Rimuovi assistito da lista"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idAssistitoListe",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "description": "ID del record AssistitiListe principale da aggiornare"
          },
          {
            "in": "query",
            "name": "stato",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "description": "Nuovo stato: 2=in assistenza, 3=completato, 4=cambio lista, 5=rinuncia, 6=annullato"
          },
          {
            "in": "query",
            "name": "azioniAltreListe",
            "required": false,
            "schema": {},
            "description": "Array di azioni per le altre liste: [{ idAssistitoListe, azione: \"mantieni\"|\"rimuovi\", statoRimozione? }]"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "string",
                      "description": "Assistito rimosso dalla lista."
                    },
                    {
                      "type": "string",
                      "description": "Success"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/update-lista-tag": {
      "put": {
        "summary": "Update lista tag",
        "description": "Aggiorna il tag di una lista esistente.",
        "tags": [
          "Update lista tag"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tag aggiornato."
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/dashboard": {
      "get": {
        "summary": "API Dashboard",
        "description": "Ritorna dati JSON per la dashboard: statistiche, stato wallet, stato Arweave, ultime operazioni.",
        "tags": [
          "API Dashboard"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Dati dashboard."
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/strutture": {
      "get": {
        "summary": "API Strutture",
        "description": "Ritorna strutture filtrabili per organizzazione.",
        "tags": [
          "API Strutture"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "organizzazione",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "description": "Filtra per ID organizzazione"
          },
          {
            "in": "query",
            "name": "id",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "description": "ID struttura per dettaglio"
          }
        ],
        "responses": {
          "200": {
            "description": "Dati strutture."
          },
          "404": {
            "description": "Struttura non trovata."
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/graph-data": {
      "get": {
        "summary": "API Graph Data",
        "description": "Ritorna tutti i dati necessari per il grafo interattivo: entita e relazioni.",
        "tags": [
          "API Graph Data"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Dati per il grafo."
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/debug": {
      "get": {
        "summary": "API Debug",
        "description": "Ritorna tutti i dati del sistema: wallet, transazioni blockchain (con decryption), DB locale, cross-references.",
        "tags": [
          "API Debug"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Dati debug completi."
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/sync-status": {
      "get": {
        "summary": "API Sync Status",
        "description": "Ritorna lo stato della sincronizzazione blockchain.",
        "tags": [
          "API Sync Status"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/sync-reset": {
      "post": {
        "summary": "Reset Sync Cache",
        "description": "Cancella la cache locale e riforza la sync dalla blockchain.",
        "tags": [
          "Reset Sync Cache"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/load-test": {
      "post": {
        "summary": "Inizializza dati di prova",
        "description": "Crea dati massivi per test UI. Solo DB locale + MAIN_DATA su blockchain.",
        "tags": [
          "Inizializza dati di prova"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/v1/wallet/get-info": {
      "get": {
        "summary": "Get wallet info",
        "description": "Restituisce le informazioni sullo stato del wallet IOTA.",
        "tags": [
          "Get wallet info"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "The requested resource"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/wallet/init": {
      "post": {
        "summary": "Init wallet",
        "description": "Inizializza il wallet IOTA 2.0 (genera mnemonic, richiede faucet).",
        "tags": [
          "Init wallet"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "The requested resource"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/wallet/reset": {
      "post": {
        "summary": "Reset wallet",
        "description": "Distrugge il wallet corrente e ne crea uno nuovo. Operazione irreversibile.",
        "tags": [
          "Reset wallet"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "The requested resource"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/public/liste": {
      "get": {
        "summary": "API Pubblico Liste",
        "description": "Ritorna tutte le liste con statistiche e coda anonimizzata. Endpoint pubblico senza autenticazione.",
        "tags": [
          "API Pubblico Liste"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Liste pubbliche con statistiche."
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/arweave/status": {
      "get": {
        "summary": "Arweave Status",
        "description": "Ritorna lo stato dettagliato di Arweave (configurazione, modo, statistiche).",
        "tags": [
          "Arweave Status"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Resource not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}